Min distance search fixed and DZT codes search draft
[*] bug fix for searching min distance for tailbiting codes.
[*] some bug fixes in classes design
[+] draft classes for searching DZT codes

+++ src/trellises/ConvCodeTrellis.java
@@ -286,7 +286,7 @@ public class ConvCodeTrellis  implements ITrellis {
 	}
 
 	@Override
-	public ITrellisIterator iterator(int layer, int vertexIndex) {
+	public ITrellisIterator iterator(int layer, long vertexIndex) {
 		return new Iterator(layer, vertexIndex);
 	}
